change videos/:id/concert/detect to concert/detect statless endpoint
