+++ backend-go/handlers/concert_handler.go
@@ -5,6 +5,7 @@ import (
 	"strconv"
 
 	"github.com/areeeeeeeb/reLive/backend-go/apperr"
+	"github.com/areeeeeeeb/reLive/backend-go/models"
 	"github.com/areeeeeeeb/reLive/backend-go/services"
 	"github.com/gin-gonic/gin"
 )
@@ -14,6 +15,7 @@ type ConcertHandler struct {
 	actService             *services.ActService
 	songPerformanceService *services.SongPerformanceService
 	videoService           *services.VideoService
+	detectionService       *services.DetectionService
 }
 
 func NewConcertHandler(
@@ -21,12 +23,14 @@ func NewConcertHandler(
 	actService *services.ActService,
 	songPerformanceService *services.SongPerformanceService,
 	videoService *services.VideoService,
+	detectionService *services.DetectionService,
 ) *ConcertHandler {
 	return &ConcertHandler{
 		concertService:         concertService,
 		actService:             actService,
 		songPerformanceService: songPerformanceService,
 		videoService:           videoService,
+		detectionService:       detectionService,
 	}
 }
 
@@ -88,6 +92,24 @@ func (h *ConcertHandler) ListSongPerformances(c *gin.Context) {
 	}
 
 	c.JSON(200, gin.H{"song_performances": result})
+}
+
+// POST /v2/api/concerts/detect
+// Stateless: returns top concert candidates for the given GPS coordinates and timestamp.
+func (h *ConcertHandler) Detect(c *gin.Context) {
+	var req models.ConcertDetectRequest
+	if err := c.ShouldBindJSON(&req); err != nil {
+		c.JSON(400, gin.H{"error": err.Error()})
+		return
+	}
+
+	result, err := h.detectionService.DetectConcert(c.Request.Context(), req)
+	if err != nil {
+		c.JSON(500, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(200, result)
 }
 
 func (h *ConcertHandler) ListVideos(c *gin.Context) {
