+++ backend-go/handlers/video_handler.go
@@ -10,12 +10,11 @@ import (
 )
 
 type VideoHandler struct {
-	videoService     *services.VideoService
-	detectionService *services.DetectionService
+	videoService *services.VideoService
 }
 
-func NewVideoHandler(videoService *services.VideoService, detectionService *services.DetectionService) *VideoHandler {
-	return &VideoHandler{videoService: videoService, detectionService: detectionService}
+func NewVideoHandler(videoService *services.VideoService) *VideoHandler {
+	return &VideoHandler{videoService: videoService}
 }
 
 // GET /videos
@@ -90,46 +89,6 @@ func (h *VideoHandler) UploadConfirm(c *gin.Context) {
 		VideoID: videoID,
 		Status:  models.VideoStatusCompleted,
 	})
-}
-
-// POST /videos/:id/concert/detect
-// Detects which concert a video belongs to using client-provided metadata.
-func (h *VideoHandler) DetectConcert(c *gin.Context) {
-	videoID, err := strconv.Atoi(c.Param("id"))
-	if err != nil {
-		c.JSON(400, gin.H{"error": "invalid video ID"})
-		return
-	}
-
-	userID := c.GetInt("user_id")
-	if userID == 0 {
-		c.JSON(401, gin.H{"error": "user not found"})
-		return
-	}
-
-	video, err := h.videoService.GetByID(c.Request.Context(), videoID)
-	if err != nil {
-		c.JSON(404, gin.H{"error": "video not found"})
-		return
-	}
-	if video.UserID != userID {
-		c.JSON(403, gin.H{"error": "forbidden"})
-		return
-	}
-
-	var req models.ConcertDetectRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
-		return
-	}
-
-	result, err := h.detectionService.DetectConcert(c.Request.Context(), videoID, req)
-	if err != nil {
-		c.JSON(500, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(200, result)
 }
 
 // DELETE /videos/:id
